% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{post_prob}
\alias{post_prob}
\title{Calculate the posterior probability that a gene is essential using MCMC.}
\usage{
post_prob(
  file_path = "insertion_density_csv_file",
  insdens_thresh = 0.025,
  init_theta = 0.05,
  niter = 21000,
  burn_in = 1000,
  acc_window = 500,
  ae_sd = 0.25,
  be_sd = 0.05,
  an_sd = 0.25,
  bn_sd = 0.05,
  print_prop_sd = T
)
}
\arguments{
\item{file_path}{string containing the path to the source csv file. The
source file must be a csv file with two columns. The first column contains
the gene name and must be called \code{gene}. The second column must contain the
insertion density and must be called \code{insdens}.}

\item{insdens_thresh}{insertion density threshold used to initialise the
essentiality status of each gene. The default value is 0.025. Use the value
that you think would a priori be your choice of insertion density threshold
to classify genes into essential and non-essential. It must be between 0 and
1.}

\item{init_theta}{initial value of the theta parameter, the prior probability that a
gene is essential. The default value is 0.05. Use your a priori estimate of the
proportion of essential genes. It must be between 0 and 1.}

\item{niter}{A positive integer greater than or equal to 2 representing the
total number of MCMC iterations including the \code{burn-in}.
The default value is 21,000.}

\item{burn_in}{A non-negative integer representing the MCMC \code{burn-in}.The default
is 1000. The \code{burn-in} samples are not included in the posterior estimates.
It must be less than niter.}

\item{acc_window}{A positive integer representing the number of steps between
updates to proposal standard deviations. The aim is to achieve an acceptance
rate of between 15\% and 35\%. The default value is 500.}

\item{ae_sd}{The proposal standard devation for the \eqn{\alpha_E} parameter.
The default value is 0.25.}

\item{be_sd}{The proposal standard devation for the \eqn{\beta_E} parameter.
The default value is 0.40.}

\item{an_sd}{The proposal standard devation for the \eqn{\alpha_N} parameter.
The default value is 0.05.}

\item{bn_sd}{The proposal standard devation for the \eqn{\beta_N} parameter.
The default value is 0.05.}

\item{print_prop_sd}{A logical value indicating whether to print the updated proposal
standard deviations and acceptance rates of each parameter. If TRUE they
will be printed  every \code{acc_window} iterations.}
}
\value{
A list of two elements. The first list element is a data frame in
 which the first column gives the gene names and the the second column gives
 the posterior probability that the gene is essential. The second list
 element is a data frame containing the posterior samples of the model
 parameters (including the burn-in). This allows the user to check
 convergence of the MCMC chain.
}
\description{
Calculate the posterior probability that a gene is essential using MCMC.
}
\examples{
sum(1,10)
}
